Update New-SlackMessage to fix pipeline
This fixes #73

What happens is that when you run via the
 pipeline, the PROCESS block is run for each
item in the `ExistingAttachements` variable.

Then it returns that item + the current one back to the pipeline.

When this happens, the next command in the pipeline starts to work on that new object.

This causes a cascade which is what causes the error you are seeing.